Fix process kill killing server
If the zeroscale process got a Ctrl-C, it would kill the server process
without warning. Now it calls the stop command.
The stop command in the Minecraft plugin didn't handle if the server was
still starting up, now it can stop while starting.
Update Status enum to not print value since it doesn't matter.