Fix conflicting read() when stopping server
Only happens when stop()ing server while it is still starting, as the
stop() call calls communicate() which reads from stdout, which is what
the await_server_start() call does. Now cancel the await_server_start()
call, and go right to shutting the server down.
Should the timeout be a parameter?