Fix socket timeout hanging server
If a socket timed out (or similar?), the code wouldn't handle that well
and would forget to stop counting that client, which would mean that the
server would never shut down.

Also, try flushing the writing pipe after writing to it, just in case it
is completely full. It really shouldn't happen, since the data we are
writing to the socket is data we just read from an identical socket, but
this might be safer.