Add check for valid client before startup
I noticed that my servers would often start up, only for them to shut
down right away after the timeout, as no one connected. It seems that my
physical server is getting port scanned often, and this would make
zeroscale think someone was trying to connect. I know this is the case
for my ssh server at least.

To combat this, I added checks to the first bits of data that the client
sends to see if it is actually a valid client. If so, it works just like
before. Otherwise, it will close the connection.

Both Minecraft and Terraria send a decent amount of info in their
opening packets, and so I use that to determine if the client is valid.

For the generic plugin, I don't do any check. If you want a valid check,
write a plugin.

I also cleaned up the regular expression checking with search() instead
of match() with greedy at the beginnings.