Fix client checking connection issues
If a client had connection issues while in the client test, bad things
would happen. Now we catch the errors better and handle them correctly.

Also add a condition in the client tests where a client sends no data at
all, and don't do stupid stuff.

Don't catch asyncio.TimoutError in the client test, as we can catch it
up above, and then log it better as a connection error.