Add pause functionality as default
Instead of stopping and starting server each time, the default method
now is to start the server on startup, then pause it. When a client
connects, unpause it, and proxy the connection right away. It's quite
transparent now. The old stopping method remains, but is not
recommended.
This is handled through signals, the same ones that a shell uses for job
sleeping.
Add a new Status, `paused`.

Move a bunch of common command line arguments into a common file. This
will be useful when another command line program is added next commit.

Abstract and refactor a bunch of ZeroScale code so that it can handle
both pausing and stopping as forms of control.

Break out a bunch of the common plugin stuff into Classes that can be
inherited, so now the Minecraft and Terraria plugins can already work
with pausing, without any changes to their classes.
The BaseServer only impliments the basic client validation from the
GenericServer.
The GenericServer impliments the basic handling of processes and the
signals to control them.
Now new plugins only have to override the methods that need changing,
like start and stop if they need to control the server through Pipes, or
simply the client validation is_valid_connection() and fake_status().