Rework command line arguments
Change the command line arguments to get rid of required options.
Default the server port to be the port that the proxy server is
listening on.
Default the server hostname to the container ID for the docker version.
This will almost always be the case anyway.
Default the listen port to 8080.
Default the server plugin to `generic`. This is what the docker version
did anyway, and most of the time this is probably what is wanted.
Rename the "server_plugin" argument to "plugin".

Add "--help" to the default docker container image command line options,
so running the container without arguments shows the usage text instead
of the "container_id missing" error message.