Add address info to debug messages and catch more errors
Try to catch exceptions thrown by the proxy. There can be errors both on
connecting to the backend (say if it is not ready for connections yet)
or from the client killing the connection. Some where being handled a
layer above, but not all.

Add address info to most debugging messages to help with debugging.